cookie + 签名

+++ server/controllers/favorite.js
@@ -14,19 +14,25 @@ module.exports = {
             data: null,
             code: ''
         }
-        let existOne  = await favoriteService.getExistOne(data)
-        if ( existOne  ) {
-            if ( existOne.mid === data.mid + '' ) {
+        let cid = ctx.cookies.get('cid')
+        let signature = ctx.cookies.get('signature')
+        if(verificationSignature(cid, signature)){
+            let existOne  = await favoriteService.getExistOne(data)
+            if ( existOne  ) {
+                if ( existOne.mid === data.mid + '' ) {
 
-                result.message = favoriteCode.FAIL_SONG_IS_EXIST
-                ctx.body = result
-                return
+                    result.message = favoriteCode.FAIL_SONG_IS_EXIST
+                    ctx.body = result
+                    return
+                }
             }
-        }
-        let addResult = await favoriteService.addFavorite( data )
-        if ( addResult && addResult.insertId * 1 > 0) {
-            result.success = true
-        } else {
+            let addResult = await favoriteService.addFavorite( data )
+            if ( addResult && addResult.insertId * 1 > 0) {
+                result.success = true
+            } else {
+                result.message = favoriteCode.ERROR_SYS
+            }
+        }else{
             result.message = favoriteCode.ERROR_SYS
         }
         ctx.body = result;
@@ -39,10 +45,16 @@ module.exports = {
             data: null,
             code: ''
         }
-        let deleteResult = await favoriteService.deleteFavorite( data )
-        if ( deleteResult && deleteResult.affectedRows == 1) {
-            result.success = true
-        } else {
+        let cid = ctx.cookies.get('cid')
+        let signature = ctx.cookies.get('signature')
+        if(verificationSignature(cid, signature)){
+            let deleteResult = await favoriteService.deleteFavorite( data )
+            if ( deleteResult && deleteResult.affectedRows == 1) {
+                result.success = true
+            } else {
+                result.message = favoriteCode.BAD_REQUEST
+            }
+        }else{
             result.message = favoriteCode.ERROR_SYS
         }
         ctx.body = result;
